requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
execnet==2.1.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.2.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
locust==2.27.0
MarkupSafe==2.1.5
msgpack==1.0.8
mysql-connector==2.2.9
outcome==1.3.0.post0
packaging==24.0
pluggy==1.5.0
psutil==5.9.8
PySocks==1.7.1
pytest~=8.2.1
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-rerunfailures==14.0
pytest-xdist==3.6.1
pyzmq==26.0.3
requests==2.31.0
selenium~=4.21.0
setuptools==69.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
zope.event==5.0
zope.interface==6.4
pydantic~=2.7.3